Board全体を保存する様に設定
